refactor: Move SleepDurationProvider to separate file
