+++ src/Dodo.HttpClient.ResiliencePolicies/RetryPolicy/IRetryPolicySettings.cs
@@ -8,9 +8,10 @@ namespace Dodo.HttpClientResiliencePolicies.RetryPolicy
 {
 	public interface IRetryPolicySettings
 	{
+		Action<DelegateResult<HttpResponseMessage>, TimeSpan> OnRetry { get; set; }
+
 		internal int RetryCount { get; }
 		internal Func<int, DelegateResult<HttpResponseMessage>, Context, TimeSpan> SleepDurationProvider { get; }
-		internal Func<DelegateResult<HttpResponseMessage>, TimeSpan, int, Context, Task> OnRetryForPolly { get; }
-		Action<DelegateResult<HttpResponseMessage>, TimeSpan> OnRetry { get; set; }
+		internal Func<DelegateResult<HttpResponseMessage>, TimeSpan, int, Context, Task> OnRetryWrapper { get; }
 	}
 }
